pyproject.toml
[tool.poetry]
name = "hotel-booking-fastapi"
version = "0.1.0"
description = ""
authors = ["Daniil Savushkin <savushkin.d@dev-system.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
alembic = "^1.13.1"
sqlalchemy = "^2.0.30"
pydantic-settings = "^2.2.1"
httpx = "^0.27.0"
sqladmin = "^0.17.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
asyncpg = "^0.29.0"
pillow = "^10.3.0"
celery = "^5.4.0"
flower = "^2.0.1"
email-validator = "^2.1.1"
bcrypt = "^4.1.3"
python-json-logger = "^2.0.7"
sentry-sdk = {extras = ["fastapi"], version = "^2.2.0"}
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
itsdangerous = "^2.2.0"
greenlet = "^3.0.3"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.4"
pytest = "^8.2.0"
black = "^24.4.2"
flake8 = "^7.0.0"
autoflake = "^2.3.1"
isort = "^5.13.2"
pyright = "^1.1.363"


[tool.poetry.group.test.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.6"
pytest-xdist = "^3.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
